Replace dapr with regular HTTP client for invoking transport company
It turns out that Dapr is unable to interact with istio correctly. The
result is that load is not spread across transport-a and transport-b.
Instead, it's routed directly to transport-a or transport-b
depending on which service registered last.

This change introduces a regular HTTP client to interact with the
transport company.